Refactor one mysql test to be able to run independently
