+++ server/datastore/datastore.go
@@ -102,5 +102,4 @@ var TestFunctions = []func(*testing.T, fleet.Datastore){
 	testSaveHostSoftware,
 	testSaveUsers,
 	testNewActivity,
-	testActivityUsernameChange,
 }
